// js/stat-evaluation-player/src/generated/PossessionTeamStats.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { LabeledFloatSums } from "./LabeledFloatSums.ts";

/**
 * Per-team accumulated possession stats.
 */
export type PossessionTeamStats = { tracked_time: number, possession_time: number, opponent_possession_time: number, neutral_time: number, labeled_time?: LabeledFloatSums, };